Restructred my file structure
