+++ api/__init__.py
@@ -6,5 +6,4 @@ sm  will be my flask instianted variable.
 """
 
 sm = Flask(__name__)
-
 from api import routes
