+++ api/routes.py
@@ -1,7 +1,15 @@
-from flask import jsonify, request,  url_for
-from . import sm, models, validators
+from flask import jsonify, json, request,  url_for
+from . import sm
+from flask import request, jsonify
+from .models.Products import Products
+from .models.sales import Sales
+from api.validators import Validators
 
-products = []
+
+"""
+These lists will store the Sales and Products.
+"""
+Products = []
 Sales = []
 
 
@@ -12,14 +20,14 @@ def add_product():
     product_id = parsejson['product_id']
     product_price = parsejson['product_price']
     product_quantity = parsejson['product_quantity']
-    products.append(
+    Products.append(
         product_name,
         product_id,
         product_price,
         product_quantity
          )
     return jsonify({'message': 'Product has been added'}), 201
-
+    
 
 @sm.route('/api/v1/products/product_id', methods=['GET'])
 def get_a_specific_product(id):
@@ -35,7 +43,7 @@ def get_all_products():
     This function returns a list of all products in the store.
     '''
     return jsonify({
-        'products': products.get_all_products()}), 200
+        'products': Products}), 200
 
 
 @sm.route('/api/v1/admin/sales', methods=['GET'])
@@ -43,7 +51,7 @@ def get_all_sales():
     '''
      Function to enable an admin get all sales records.
     '''
-    return jsonify({'sales': Sales.get_all_sales}), 200
+    return jsonify({'Sales': Sales, 'msg': 'hey'}), 200
 
 
 @sm.route('/api/v1/attendant/sales', methods=['POST'])
@@ -52,14 +60,14 @@ def add_sale():
     This function enables a store attendant to add sale
     order to the list of orders.
     '''
-    parsejson = request.get_json()
-    product_name = parsejson['product_name']
-    product_category = parsejson['product_category']
-    quantity = parsejson['quantity']
-    product_price = parsejson['product_price']
-    date_added = parsejson['date_added']
-    product_id = parsejson['product_id']
-    Attendant_name = parsejson['attendant_name']
+    data = request.get_json()
+    product_name = data['product_name']
+    product_category = data['product_category']
+    quantity = data['quantity']
+    product_price = data['product_price']
+    date_added = data['date_added']
+    product_id = data['product_id']
+    Attendant_name = data['attendant_name']
     Sales.append(
         product_name,
         product_category,
