+++ run.py
@@ -1,5 +1,6 @@
 from api import sm
 
+
 """ run the application """
 if __name__ == '__main__':
     sm.run(debug=True)
